api/requirements.txt
flask
google-cloud-storage
google-cloud-bigquery
google-cloud-aiplatform
pandas